refactor(examples): separate capture runtime context
